Problems issues #3 are fixed. References must be made.
