+++ library/AkaiKKRTestScript/akaikkr_testscript/__init__.py
@@ -2,7 +2,7 @@
 # Copyright (c) 2021 AkaiKKRteam.
 # Distributed under the terms of the Apache License, Version 2.0.
 
-__version__ = '2021.12.08'
+__version__ = '2021.12.09'
 from .testrun_class import *
 from .exeutil import *
 from .resultutil import *
